pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]  
include = ["satop_platform*"]

[project.scripts]
satop-api = "satop_platform:main.main"
satop-cli = "satop_platform:cli.cli_app"

[project]
name = "satop_platform"
version = "0.1.0"
authors = [
    { name = "Aleksander", email = "202004314@post.au.dk" },
    { name = "Tobias", email = "202004621@post.au.dk" }
]

requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.12",
    "networkx>=3.5",
    "PyJWT>=2.10",
    "python-dotenv>=1.0.1",
    "python-multipart>=0.0.19",
    "PyYAML>=6.0.2",
    "sqlmodel>=0.0.24",
    "typer>=0.16",
    "uvicorn>=0.34",
    "websockets>=15.0.1"
]

[project.optional-dependencies]
test = [
    "pytest",
    "pytest-asyncio",
    "httpx",
    "PyYAML"
]

lint = [
    "ruff",
    "black",
    "mypy"
]

[tool.ruff]
line-length = 88
src = ["satop_platform", "satop_plugins", "test"]
select = ["E", "F", "W", "I"]

[tool.black]
line-length = 88

[tool.mypy]
strict = true
ignore_missing_imports = true
plugins = ["pydantic.mypy"]